requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
CacheControl==0.12.10
cachetools==4.2.4
cachy==0.3.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.8
cleo==0.8.1
click==8.0.1
clikit==0.6.2
coverage==6.2
crashtest==0.3.1
cryptography==3.4.8
distlib==0.3.3
dnspython==2.1.0
dynaconf==3.1.7
email-validator==1.1.3
factory-boy==3.2.0
Faker==9.5.2
fastapi==0.68.1
filelock==3.4.0
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.2.1
google-cloud-logging==2.7.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
h11==0.12.0
html5lib==1.1
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.2
keyring==21.8.0
lockfile==0.12.2
MarkupSafe==2.0.1
motor==2.5.1
msgpack==1.0.3
packaging==21.2
passlib==1.7.4
pastel==0.2.1
pexpect==4.8.0
pip==21.2.3
pkginfo==1.8.1
platformdirs==2.4.0
pluggy==1.0.0
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.3.0
pylev==1.4.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.8
setuptools==57.4.0
shellingham==1.4.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.9
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
tomlkit==0.7.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.10.0
webencodings==0.5.1